fix: add connect timeout and bypass connect_urls
Use a optional configurable timeout when connecting to a server address.

Add a hook to NOT cache connect_urls returned by a Server's INFO message
as this means we may retry connecting to bogus IP addresses.
This does not mean we never retry connecting to bogus IP addresses as
we still try to connect to whatever DNS tells us to, which is not always
up to date.